fixes after merging

+++ src/main/java/com/app_gestion_formation/application_gestion_formation/repositories/FormationRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app_gestion_formation.application_gestion_formation.models.Formation;
 
-public interface FormationRepo extends JpaRepository <Formation, Long> {
+public interface FormationRepo extends JpaRepository <Formation, Integer> {
     boolean existsById(Integer id);
 
 }
