+++ src/main/java/com/app_gestion_formation/application_gestion_formation/repositories/ParticipantRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.app_gestion_formation.application_gestion_formation.models.Participant;
 
 public interface ParticipantRepo extends JpaRepository<Participant, Long>  {
-    public  Participant  findById(long id);
+    public  Participant  findById(int id);
 	boolean existsByEmail(String email);
 }
